Backend/requirements.txt
fastapi
pydantic
nest-asyncio
uvicorn
pymongo
pandas
python-dotenv
pyngrok
python-multipart
motor
cryptography
python-jose[cryptography]
datetime
passlib[bcrypt]